requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via django
billiard==4.2.2
    # via celery
celery==5.5.3
    # via django-celery-beat
certifi==2025.11.12
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
cron-descriptor==2.0.6
    # via django-celery-beat
distlib==0.4.0
    # via virtualenv
django==5.2.7
    # via
    #   django-celery-beat
    #   django-crontab
    #   django-jazzmin
    #   django-timezone-field
    #   djangorestframework
django-celery-beat==2.8.1
django-crontab==0.7.1
django-jazzmin==3.0.2
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.16.1
filelock==3.20.0
    # via virtualenv
icalendar==6.3.2
identify==2.6.15
    # via pre-commit
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
kombu==5.5.4
    # via celery
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   kombu
    #   pytest
platformdirs==4.5.0
    # via virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.3.0
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via pytest-django
pytest-django==4.11.1
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   icalendar
python-dotenv==1.2.1
pyyaml==6.0.3
    # via pre-commit
redis==7.0.1
requests==2.32.5
ruff==0.14.3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via cron-descriptor
tzdata==2025.2
    # via
    #   django
    #   django-celery-beat
    #   icalendar
    #   kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.35.4
    # via pre-commit
wcwidth==0.2.14
    # via prompt-toolkit
